Refactoring the answer property of the question
Before this PR, the answer of a question was an ExternalUserId (eg,
Slack user id).

This design led to some wierd code in the upper layer and is not very
future proof nor easy to understand.

So this PR introduces a proper 'Answer' object which has a
ExternalUserId property (it wraps an external user id really.).

All the code is refactored accordingly.