[Refactoring] - application layer
In order to take into account the passing case of the quiz is completed
by the user. This information is not passed on via exceptions nor
guessed from the repository because there is not questions to answer
(which could be an actual error).

In order to do this, each application service now returns a custom
response object with properties:
- isQuizCompleted boolean
- and other any data the handler wants to return